refactor(codex): split app-server event projector
